+++ packages/alfabit-button/src/Button.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import { defaultTheme, Theme } from '@gabriel_veroneze/alfabit-tokens'
 import { StyledButton } from './Button.styles'
 
 export interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
@@ -9,6 +10,7 @@ export interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElemen
     iconPosition?: 'left' | 'right'
     disabled?: boolean
     onClick?: () => void
+    theme?: Theme
 }
 
 export const Button = ({
@@ -17,6 +19,7 @@ export const Button = ({
     children,
     disabled,
     onClick,
+    theme = defaultTheme,
     ...props
 }: ButtonProps) => {
     return (
@@ -25,6 +28,7 @@ export const Button = ({
             size={size}
             disabled={disabled}
             onClick={onClick}
+            theme={theme}
             {...props}
         >
             {children}