+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # apps
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+    - @gabriel_veroneze/alfabit-button@1.1.0
+    - @gabriel_veroneze/alfabit-tokens@1.1.0
+    - @gabriel_veroneze/alfabit-textblock@1.0.0
+
 ## 0.0.2
 
 ### Patch Changes